refactor(database): remove redundant toint() for keyversion
- Removed unnecessary `.toInt()` calls for `keyVersion` field in `EncryptedSecretLocalDaoImpl`.
- Aligns `keyVersion` mapping with updated schema/data class type.

+++ app/src/commonMain/kotlin/eu/torvian/chatbot/app/database/dao/EncryptedSecretLocalDaoImpl.kt
@@ -50,7 +50,7 @@ class EncryptedSecretLocalDaoImpl(
                 id = row.id,
                 encryptedSecret = row.encryptedSecret,
                 encryptedDEK = row.encryptedDEK,
-                keyVersion = row.keyVersion.toInt(),
+                keyVersion = row.keyVersion,
                 createdAt = row.createdAt,
                 updatedAt = row.updatedAt
             )
@@ -87,7 +87,7 @@ class EncryptedSecretLocalDaoImpl(
                             id = row.id,
                             encryptedSecret = row.encryptedSecret,
                             encryptedDEK = row.encryptedDEK,
-                            keyVersion = row.keyVersion.toInt(),
+                            keyVersion = row.keyVersion,
                             createdAt = row.createdAt,
                             updatedAt = row.updatedAt
                         )
@@ -120,7 +120,7 @@ class EncryptedSecretLocalDaoImpl(
                     id = row.id,
                     encryptedSecret = row.encryptedSecret,
                     encryptedDEK = row.encryptedDEK,
-                    keyVersion = row.keyVersion.toInt(),
+                    keyVersion = row.keyVersion,
                     createdAt = row.createdAt,
                     updatedAt = row.updatedAt
                 )
